# client-sdks/typescript/private.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sema4.ai Agent Server Private API Version 2",
    "version": "2.0.2-alpha"
  },
  "servers": [
    {
      "url": "/api/v2"
    }
  ],
  "paths": {
...
    "/capabilities/providers": {
      "get": {
        "tags": ["capabilities"],
        "summary": "Get Providers",
        "description": "This endpoint returns detailed information about all supported model platforms\nand their configuration parameters, including field descriptions, types,\nand whether they are required or optional.\n\nThe schema returned represents a union of all possible platform parameter\ntypes.",
        "operationId": "get_providers_capabilities_providers_get",
        "parameters": [
          {
            "name": "response_model",
            "in": "query",
            "required": false,
            "schema": {
              "title": "Response Model"
            }
          },
          {
            "name": "summary",
            "in": "query",
            "required": false,
            "schema": {
              "default": "Get OpenAPI Schema for Provider Parameters",
              "title": "Summary"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/capabilities/providers/{kind}/test": {
      "post": {
        "tags": ["capabilities"],
        "summary": "Test Model Platform Params",
        "description": "This endpoint is used to test platform parameters by making a series\nof basic requests to the platform.\n\nThe endpoint returns a dictionary with the results of the tests.",
        "operationId": "test_model_platform_params_capabilities_providers__kind__test_post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Platform Params"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}